# likenft-indexer/openapi/schema.yaml
openapi: 3.0.3
info:
  title: LikeNFT Indexer API
  version: 0.1.0
paths:
...
  /booknft/{id}:
    get:
      tags:
        - BookNFT
      summary: Query BookNFT
      description: Query BookNFT
      operationId: bookNFT
      parameters:
        - name: id
          in: path
          description: evm address of the BookNFT
          required: true
          schema:
            type: string
      responses:
        "200":
          description: BookNFT with requested ID was found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookNFT"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
        cosmos_address:
          type: string
        evm_address:
          type: string
        likeid:
          type: string
      required:
        - id
        - evm_address
...
    BookNFT:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
        name:
          type: string
        symbol:
          type: string
        owner_address:
          type: string
        minter_addresses:
          type: array
          items:
            type: string
        updater_addresses:
          type: array
          items:
            type: string
        total_supply:
          $ref: "#/components/schemas/BigInt"
        max_supply:
          $ref: "#/components/schemas/Uint64"
        metadata:
          $ref: "#/components/schemas/ContractLevelMetadata"
        banner_image:
          type: string
        featured_image:
          type: string
        deployed_block_number:
          $ref: "#/components/schemas/Uint64"
        minted_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/Account"
      required:
        - id
        - address
        - name
        - symbol
        - minter_addresses
        - updater_addresses
        - total_supply
        - max_supply
        - banner_image
        - featured_image
        - deployed_block_number
        - minted_at
        - updated_at
...
    BigInt:
      type: string
      example: "1"
    ContractLevelMetadata:
      type: object
      properties:
        name:
          type: string
        symbol:
          type: string
        description:
          type: string
        image:
          type: string
        banner_image:
          type: string
        featured_image:
          type: string
        external_link:
          type: string
        collaborators:
          type: array
          items:
            type: string
      additionalProperties: true
...
    Uint64:
      type: string
      example: "1"
    Error:
      description: Represents error object
      type: object
      properties:
        code:
          type: integer
          format: int64
          default: 500
        message:
          type: string
        sentry_error_id:
          type: string
      required:
        - code
        - message